+++ src/main/java/org/curtinfrc/frc2026/Robot.java
@@ -17,9 +17,9 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
-import org.curtinfrc.frc2026.Subsystem.Intake.Intake;
-import org.curtinfrc.frc2026.Subsystem.Intake.IntakeIODev;
-import org.curtinfrc.frc2026.Subsystem.Intake.IntakeIOSim;
+import org.curtinfrc.frc2026.Subsystem.Intake;
+import org.curtinfrc.frc2026.Subsystem.IntakeIODev;
+import org.curtinfrc.frc2026.Subsystem.IntakeIOSim;
 import org.curtinfrc.frc2026.drive.Drive;
 import org.curtinfrc.frc2026.drive.GyroIO;
 import org.curtinfrc.frc2026.drive.GyroIOPigeon2;
