+++ src/main/java/org/curtinfrc/frc2026/Subsystem/IntakeIOComp.java
@@ -1,3 +1,3 @@
-package org.curtinfrc.frc2026.Subsystem.Intake;
+package org.curtinfrc.frc2026.Subsystem;
 
 public class IntakeIOComp implements IntakeIO {}
